Initialize s:placeholder_texts (#11)
While pressing a Tab in a  file that contains `{{++}}` it would error out with
the error below. I actually noticed this while editing minisnip's documentation.

The problem is that there isn't a snippet we ever completed.

This fixes the error by initializing `s:placeholder_texts` and the `'<` mark. It
also adds `abort` to all the functions so it will exit in the first error
instead if continuing on errors (which will usually just spawn more errors, like
in this case).

---

	Error detected while processing function minisnip#Minisnip:
	line   19:
	E20: Mark not set
	Error detected while processing function minisnip#Minisnip[23]..<SNR>68_SelectPlaceho
	lder:
	line   22:
	E121: Undefined variable: s:placeholder_texts
	E116: Invalid arguments for function add
	2 fewer lines; before #46  18 seconds ago
	29 lines yanked into "s